[CHORE]: FFmpeg 실행기 폴더 변경
